wip: read conf
