+++ pkg/bridge/ai/api_server.go
@@ -3,19 +3,13 @@ package ai
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"net"
 	"net/http"
-	"strings"
 	"sync"
 	"time"
 
-	"gopkg.in/yaml.v3"
-
 	gonanoid "github.com/matoous/go-nanoid/v2"
 	"github.com/yomorun/yomo/ai"
-	"github.com/yomorun/yomo/core"
-	"github.com/yomorun/yomo/core/frame"
 	"github.com/yomorun/yomo/core/ylog"
 )
 
@@ -24,11 +18,6 @@ const (
 )
 
 var (
-	ErrNotExistsProvider     = errors.New("llm provider does not exist")
-	ErrNotImplementedService = errors.New("llm service is not implemented")
-	ErrConfigNotFound        = errors.New("ai config was not found")
-	ErrConfigFormatError     = errors.New("ai config format is incorrect")
-
 	// RequestTimeout is the timeout for the request, default is 5 seconds
 	RequestTimeout = 5 * time.Second
 )
@@ -43,18 +32,34 @@ type BasicAPIServer struct {
 	*Config
 	// ZipperAddr is the address of the zipper
 	ZipperAddr string
-	// Provider is the AI provider
-	Provider AIProvider
+	// Provider is the llm provider
+	Provider LLMProvider
+	// serviceCredential is the credential when using the service
+	serviceCredential string
+}
+
+// Serve starts the Basic API Server
+func Serve(config *Config, zipperListenAddr string, credential string) error {
+	provider, err := GetProviderAndSetDefault(config.Server.Provider)
+	if err != nil {
+		return err
+	}
+	srv, err := NewBasicAPIServer(provider.Name(), config, zipperListenAddr, provider, credential)
+	if err != nil {
+		return err
+	}
+	return srv.Serve()
 }
 
 // NewBasicAPIServer creates a new restful service
-func NewBasicAPIServer(name string, config *Config, zipperAddr string, provider AIProvider) (*BasicAPIServer, error) {
+func NewBasicAPIServer(name string, config *Config, zipperAddr string, provider LLMProvider, credential string) (*BasicAPIServer, error) {
 	zipperAddr = parseZipperAddr(zipperAddr)
 	return &BasicAPIServer{
-		Name:       name,
-		Config:     config,
-		ZipperAddr: zipperAddr,
-		Provider:   provider,
+		Name:              name,
+		Config:            config,
+		ZipperAddr:        zipperAddr,
+		Provider:          provider,
+		serviceCredential: credential,
 	}, nil
 }
 
@@ -65,16 +70,17 @@ func NewBasicAPIServer(name string, config *Config, zipperAddr string, provider 
 // returned as the response to the user's question.
 func (a *BasicAPIServer) Serve() error {
 	handler := http.NewServeMux()
+	// get service instance
+	service, err := NewService(a.serviceCredential, a.ZipperAddr, a.Provider)
+	if err != nil {
+		ylog.Error("Start BasicAPIServer failed", "err", err)
+		return err
+	}
+
 	// GET /overview
 	handler.HandleFunc("/overview", func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
-		credential := getBearerToken(r)
-		service, err := NewService(credential, a.ZipperAddr, a.Provider)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
-			return
-		}
+		// credential := getBearerToken(r)
 		resp, err := service.GetOverview()
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
@@ -149,13 +155,13 @@ func (a *BasicAPIServer) Serve() error {
 	handler.HandleFunc("/invoke", func(w http.ResponseWriter, r *http.Request) {
 		// log := ylog.With("path", pattern, "method", r.Method)
 		// get bearer token from request, it's yomo credential
-		credential := getBearerToken(r)
-		service, err := NewService(credential, a.ZipperAddr, a.Provider)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
-			return
-		}
+		// credential := getBearerToken(r)
+		// service, err := NewService(credential, a.ZipperAddr, a.Provider)
+		// if err != nil {
+		// 	w.WriteHeader(http.StatusInternalServerError)
+		// 	json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
+		// 	return
+		// }
 
 		defer r.Body.Close()
 		reqID, err := gonanoid.New(6)
@@ -270,230 +276,6 @@ func (a *BasicAPIServer) Serve() error {
 	return httpServer.ListenAndServe()
 }
 
-// ======================= Package Functions =======================
-// Serve starts the Basic API Server
-func Serve(config *Config, zipperListenAddr string) error {
-	provider, err := GetProviderAndSetDefault(config.Server.Provider)
-	if err != nil {
-		return err
-	}
-	srv, err := NewBasicAPIServer(provider.Name(), config, zipperListenAddr, provider)
-	if err != nil {
-		return err
-	}
-	return srv.Serve()
-}
-
-// RegisterFunction registers the AI function
-func RegisterFunction(tag uint32, functionDefinition []byte, connID string) error {
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		return err
-	}
-	fd := ai.FunctionDefinition{}
-	err = json.Unmarshal(functionDefinition, &fd)
-	if err != nil {
-		ylog.Error("unmarshal function definition", "error", err)
-		return err
-	}
-	ylog.Info("register function", "connID", connID, "name", fd.Name, "tag", tag, "function", string(functionDefinition))
-	return provider.RegisterFunction(tag, &fd, connID)
-}
-
-// UnregisterFunction unregister the AI function
-func UnregisterFunction(name string, connID string) error {
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		return err
-	}
-	return provider.UnregisterFunction(name, connID)
-}
-
-// ListToolCalls lists the AI tool calls
-func ListToolCalls() (map[uint32]ai.ToolCall, error) {
-	provider, err := GetDefaultProvider()
-	if err != nil {
-		return nil, err
-	}
-	return provider.ListToolCalls()
-}
-
-// func GetChatCompletions(appID string, prompt string) (*ai.ChatCompletionsResponse, error) {
-// 	provider, err := GetDefaultProvider()
-// 	if err != nil {
-// 		return nil, err
-// 	}
-// 	service, ok := provider.(AIService)
-// 	if !ok {
-// 		return nil, ErrNotImplementedService
-// 	}
-// 	return service.GetChatCompletions(appID, prompt)
-// }
-
-// ConnMiddleware returns a ConnMiddleware that can be used to intercept the connection.
-func ConnMiddleware(next core.ConnHandler) core.ConnHandler {
-	return func(conn *core.Connection) {
-		// check sfn type and is ai function
-		if conn.ClientType() != core.ClientTypeStreamFunction {
-			next(conn)
-			return
-		}
-		for {
-			f, err := conn.FrameConn().ReadFrame()
-			// unregister ai function on any error
-			if err != nil {
-				conn.Logger.Error("failed to read frame on ai middleware", "err", err)
-				conn.Logger.Info("error type", "type", fmt.Sprintf("%T", err))
-				// appID, _ := conn.Metadata().Get(metadata.AppIDKey)
-				name := conn.Name()
-				conn.Logger.Info("unregister ai function", "name", name, "connID", conn.ID())
-				UnregisterFunction(name, conn.ID())
-				return
-			}
-			if ff, ok := f.(*frame.AIRegisterFunctionFrame); ok {
-				err := conn.FrameConn().WriteFrame(&frame.AIRegisterFunctionAckFrame{AppID: ff.AppID, Tag: ff.Tag})
-				if err != nil {
-					conn.Logger.Error("failed to write ai RegisterFunctionAckFrame", "app_id", ff.AppID, "tag", ff.Tag, "err", err)
-					return
-				}
-				// register ai function
-				err = RegisterFunction(ff.Tag, ff.Definition, conn.ID())
-				if err != nil {
-					conn.Logger.Error("failed to register ai function", "app_id", ff.AppID, "tag", ff.Tag, "err", err)
-					return
-				}
-				conn.Logger.Info("register ai function success", "app_id", ff.AppID, "tag", ff.Tag, "definition", string(ff.Definition))
-			}
-			next(conn)
-		}
-	}
-}
-
-// server:
-//   host: http://localhost
-//   port: 8000
-//   endpoints:
-//     chat_completions: /chat/completions
-//   credential: token:<CREDENTIAL>
-//   provider: azopenai
-//
-// providers:
-//   azopenai:
-//     api_key:
-//     api_endpoint:
-//
-//   openai:
-//     api_key:
-//     api_endpoint:
-//
-//   huggingface:
-//     model:
-
-// Config is the configuration of AI bridge
-type Config struct {
-	Server    Server              `yaml:"server"`
-	Providers map[string]Provider `yaml:"providers"`
-}
-
-// Server is the configuration of server
-type Server struct {
-	Addr string `yaml:"addr"`
-	// Endpoints map[string]string `yaml:"endpoints"`
-	Provider string `yaml:"provider"`
-}
-
-// Provider is the configuration of AI provider
-type Provider = map[string]string
-
-// map[ai:
-//	map[providers:
-//		map[azopenai:
-//			map[api_endpoint:<nil>
-//					api_key:<nil>]
-//				huggingface:map[model:<nil>]
-//				openai:map[api_endpoint:<nil> api_key:<nil>]]
-//	server:map[
-//		credential:token:<CREDENTIAL>
-//		endpoints:map[chat_completions:/chat/completions]
-//		host:http://localhost
-//		port:8000
-//		provider:azopenai]]]
-
-// ParseConfig parses the AI config from conf
-func ParseConfig(conf map[string]any) (config *Config, err error) {
-	section, ok := conf["ai"]
-	if !ok {
-		err = ErrConfigNotFound
-		return
-	}
-	aiConfig, ok := section.(map[string]any)
-	if !ok {
-		err = ErrConfigFormatError
-		return
-	}
-	data, e := yaml.Marshal(aiConfig)
-	if e != nil {
-		err = e
-		ylog.Error("marshal ai config", "err", err.Error())
-		return
-	}
-	err = yaml.Unmarshal(data, &config)
-	if err != nil {
-		ylog.Error("unmarshal ai config", "err", err.Error())
-		return
-	}
-	// defaults values
-	if config.Server.Addr == "" {
-		config.Server.Addr = ":8000"
-	}
-	// // endpoints
-	// if config.Server.Endpoints == nil {
-	// 	config.Server.Endpoints = map[string]string{
-	// 		"chat_completions": DefaultChatCompletionsEndpoint,
-	// 	}
-	// }
-	ylog.Info("parse AI config success")
-	return
-}
-
-// parseZipperAddr parses the zipper address from zipper listen address
-func parseZipperAddr(addr string) string {
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		ylog.Error("invalid zipper address, return default",
-			"addr", addr,
-			"default", DefaultZipperAddr,
-			"err", err.Error(),
-		)
-		return DefaultZipperAddr
-	}
-	if host == "localhost" {
-		return addr
-	}
-	ip := net.ParseIP(host)
-	if ip == nil {
-		ylog.Error("invalid zipper address, return default",
-			"addr", addr,
-			"default", DefaultZipperAddr,
-		)
-		return DefaultZipperAddr
-	}
-	if !ip.IsUnspecified() {
-		addr = ip.String() + ":" + port
-		// ylog.Info("parse zipper address", "addr", addr)
-		return addr
-	}
-	localIP, err := getLocalIP()
-	if err != nil {
-		ylog.Error("get local ip, return default",
-			"default", DefaultZipperAddr,
-			"err", err.Error(),
-		)
-		return DefaultZipperAddr
-	}
-	return localIP + ":" + port
-}
-
 func getLocalIP() (string, error) {
 	addrs, err := net.InterfaceAddrs()
 	if err != nil {
@@ -510,18 +292,18 @@ func getLocalIP() (string, error) {
 	return "", errors.New("not found local ip")
 }
 
-// getBearerToken returns the bearer token from the request
-func getBearerToken(req *http.Request) string {
-	auth := req.Header.Get("Authorization")
-	if auth == "" {
-		return ""
-	}
-	if !strings.HasPrefix(auth, "Bearer") {
-		ylog.Error("invalid Authorization header", "header", auth)
-		return ""
-	}
-	token := strings.TrimPrefix(auth, "Bearer ")
-	return token
-	// TEST: only test
-	// return "token:Happy New Year"
-}
+// // getBearerToken returns the bearer token from the request
+// func getBearerToken(req *http.Request) string {
+// 	auth := req.Header.Get("Authorization")
+// 	if auth == "" {
+// 		return ""
+// 	}
+// 	if !strings.HasPrefix(auth, "Bearer") {
+// 		ylog.Error("invalid Authorization header", "header", auth)
+// 		return ""
+// 	}
+// 	token := strings.TrimPrefix(auth, "Bearer ")
+// 	return token
+// 	// TEST: only test
+// 	// return "token:Happy New Year"
+// }
